Fix for http://fusesource.com/issues/browse/RMIVIAJMS-4 as well as some timing issues and CGLIB issues.
Timing issue: When object were exported via ExplicitDestinationSkeleton (or otherwise) we kick of a receive 
thread that asynchronously creates the consumer. If an immediate calls i made to the returned proxy prior to the
receiver being create the invocation will be missed.

CGLib: Changing CGLib to intercept Object method calls prior to this fix things like finalize were being proxied.